SendCryptoDoneView refactored
